refactor: usePin 분리
